backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
openai-whisper
torch
torchaudio
ffmpeg-python
gunicorn==23.0.0
